# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CortexDB Beta API
  version: 0.2.0-beta.2
  description: 'Beta HTTP API for the local single-node CortexDB
    engine. The HTTP layer is async, while the database core is a local blocking engine
    executed through a blocking task boundary.

    '
servers:
- url: http://127.0.0.1:8181
paths:
...
  /v1/validate:
    get:
      summary: Validate storage integrity
      parameters:
      - $ref: '#/components/parameters/Tenant'
      responses:
        '200':
          description: Validation report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
...
  /v1/search:
    post:
      summary: Keyword or vector search
      description: |
        **Alpha/Experimental.** Search quality, HNSW graph behavior, and ANN fallback
        semantics are under active development. Exact vector scan is the safe default;
        ANN may fall back to exact when the graph is empty or under-returning;
        fallback and SLO guardrails are reported in ann_report.
      parameters:
      - $ref: '#/components/parameters/Tenant'
      - $ref: '#/components/parameters/Scope'
      - name: mode
        in: query
        schema:
          type: string
          enum:
          - keyword
          - vector
          - hybrid
          - auto
      - name: algorithm
        in: query
        description: Vector search algorithm. `ann` may fall back to exact scan when
          no graph is available.
        schema:
          type: string
          enum:
          - ann
          - exact
      - name: q
        in: query
        schema:
          type: string
      - name: vector
        in: query
        schema:
          type: string
      - name: embed_query
        in: query
        description: Generate vector from `q` through the server embedding config when `vector` is omitted.
        schema:
          type: boolean
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
      - name: fallback
        in: query
        description: Enable ANN fallback to exact scan when ANN is unavailable or underperforming.
        schema:
          type: boolean
      - name: min_recall
        in: query
        description: Minimum acceptable ANN recall (fraction like 0.75 or percent like 75).
        schema:
          type: string
      - name: fallback_scan_cap
        in: query
        description: Max number of candidates exact scan may use when fallback is needed.
        schema:
          type: integer
          minimum: 0
      - name: max_visited_candidates
        in: query
        description: Maximum number of HNSW nodes to visit before forcing ANN fallback.
        schema:
          type: integer
          minimum: 0
      - name: require_slo
        in: query
        description: Enable strict ANN/HNSW SLO reporting. Responses include production_safe=false when recall, graph, fallback, or visit-budget guardrails are violated.
        schema:
          type: boolean
      - name: no_fallback_rollout
        in: query
        description: Explicitly evaluate the selected ANN profile for fallback-free HNSW serving. This does not disable global exact fallback by itself.
        schema:
          type: boolean
      - name: no_fallback_min_recall
        in: query
        description: Optional fallback-free rollout recall threshold. Requires no_fallback_rollout=true.
        schema:
          type: string
      - name: no_fallback_profile
        in: query
        description: Use the persisted operator-managed rollout profile when set to active or true. Cannot be combined with no_fallback_rollout.
        schema:
          type: string
          enum:
          - active
          - 'true'
          - 'false'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400': *id001
        '500': *id002
...
  /v1/search/ann-evaluate:
    post:
      summary: Compare persisted ANN search against exact vector scan
      parameters:
      - $ref: '#/components/parameters/Tenant'
      - $ref: '#/components/parameters/Scope'
      - name: vector
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
      - name: fallback
        in: query
        description: Enable ANN fallback to exact scan when ANN is unavailable or underperforming.
        schema:
          type: boolean
      - name: min_recall
        in: query
        description: Minimum acceptable ANN recall (fraction like 0.75 or percent like 75).
        schema:
          type: string
      - name: fallback_scan_cap
        in: query
        description: Max number of candidates exact scan may use when fallback is needed.
        schema:
          type: integer
          minimum: 0
      - name: max_visited_candidates
        in: query
        description: Maximum number of HNSW nodes to visit before forcing ANN fallback.
        schema:
          type: integer
          minimum: 0
      - name: require_slo
        in: query
        description: Enable strict ANN/HNSW SLO reporting. Responses include production_safe=false when recall, graph, fallback, or visit-budget guardrails are violated.
        schema:
          type: boolean
      - name: no_fallback_rollout
        in: query
        description: Explicitly evaluate the selected ANN profile for fallback-free HNSW serving. This does not disable global exact fallback by itself.
        schema:
          type: boolean
      - name: no_fallback_min_recall
        in: query
        description: Optional fallback-free rollout recall threshold. Requires no_fallback_rollout=true.
        schema:
          type: string
      - name: no_fallback_profile
        in: query
        description: Use the persisted operator-managed rollout profile when set to active or true. Cannot be combined with no_fallback_rollout.
        schema:
          type: string
          enum:
          - active
          - 'true'
          - 'false'
      responses:
        '200':
          description: ANN recall evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnEvaluationResponse'
        '400': *id001
        '500': *id002
...
components:
  parameters:
    Tenant:
      name: tenant
      in: query
      required: false
      description: Optional database realm. Omit or use `default` for the root database.
      schema:
        type: string
...
    Scope:
      name: scope
      in: query
      required: true
      schema:
        type: string
...
  schemas:
...
    ValidationResponse:
      type: object
      required:
      - ok
      - manifest_ok
      - wal_ok
      - errors
      properties:
        ok:
          type: boolean
        manifest_ok:
          type: boolean
        wal_ok:
          type: boolean
        live_segments_checked:
          type: integer
        bitmap_indexes_checked:
          type: integer
        lexical_indexes_checked:
          type: integer
        vector_indexes_checked:
          type: integer
        hnsw_graphs_checked:
          type: integer
        cells_checked:
          type: integer
        wal_records_checked:
          type: integer
          format: uint64
        wal_safe_truncate_offset:
          type: integer
          format: uint64
        errors:
          type: array
          items:
            type: string
...
    SearchRoutingDecision:
      type: object
      required:
      - requested_mode
      - selected_strategy
      - reason
      - text_available
      - vector_available
      properties:
        requested_mode:
          type: string
        selected_strategy:
          type: string
          enum:
          - keyword
          - vector_exact
          - vector_ann
          - hybrid
        reason:
          type: string
        text_available:
          type: boolean
        vector_available:
          type: boolean
    SearchResponse:
      type: object
      required:
      - search_mode
      - ann_report
      - results
      properties:
        search_mode:
          type: string
          enum:
          - keyword
          - vector_exact
          - vector_ann
          - hybrid
        routing:
          anyOf:
          - $ref: '#/components/schemas/SearchRoutingDecision'
          - type: 'null'
        rerank:
          type: string
          enum:
          - weighted
          description: Present when request query parameter rerank=weighted enabled weighted reranking.
        ann_report:
          anyOf:
          - $ref: '#/components/schemas/AnnSearchReport'
          - type: 'null'
        no_fallback_decision:
          anyOf:
          - $ref: '#/components/schemas/AnnNoFallbackDecision'
          - type: 'null'
        results:
          type: array
          items:
            type: object
            required:
            - cell_id
            - score
            - lexical_score
            - vector_score
            - payload
            properties:
              cell_id:
                type: integer
                format: uint64
              score:
                type: integer
                format: uint64
              lexical_score:
                type: integer
                format: uint64
              vector_score:
                type: integer
                format: uint64
              payload:
                type: string
    AnnSearchReport:
      type: object
      required:
      - path
      - fallback_reason
      - fallback_performed
      - requested_limit
      - allowed_candidates
      - graph_nodes
      - returned_candidates
      - visited_candidates
      - recall_q16
      - min_recall_q16
      - hnsw_max_neighbors
      - hnsw_ef_search
      - hnsw_ef_construction
      - hnsw_layer_count
      - upper_graph_edges
      - require_slo
      - production_safe
      - slo_violations
      properties:
        path:
          type: string
          enum:
          - hnsw_graph
          - exact_fallback
        fallback_reason:
          anyOf:
          - type: string
            enum:
            - empty_graph
            - invalid_graph
            - stale_graph
            - insufficient_results
            - low_recall
            - visit_budget_exceeded
            - no_persisted_segments
            - uncheckpointed_changes
            - hnsw_disabled
          - type: 'null'
        fallback_performed:
          type: boolean
        requested_limit:
          type: integer
        allowed_candidates:
          type: integer
        graph_nodes:
          type: integer
        returned_candidates:
          type: integer
        visited_candidates:
          type: integer
          minimum: 0
        max_visited_candidates:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        recall_q16:
          anyOf:
          - type: integer
            minimum: 0
            maximum: 65535
          - type: 'null'
        min_recall_q16:
          anyOf:
          - type: integer
            minimum: 0
            maximum: 65535
          - type: 'null'
        hnsw_max_neighbors:
          type: integer
          minimum: 0
        hnsw_ef_search:
          type: integer
          minimum: 0
        hnsw_ef_construction:
          type: integer
          minimum: 0
          description: HNSW construction beam used when the persisted graph was built.
        hnsw_layer_count:
          type: integer
          minimum: 1
        upper_graph_edges:
          type: integer
          minimum: 0
        require_slo:
          type: boolean
        production_safe:
          type: boolean
        slo_violations:
          type: array
          items:
            type: string
            enum:
            - empty_graph
            - invalid_graph
            - insufficient_results
            - low_recall
            - visit_budget_exceeded
            - no_persisted_segments
            - uncheckpointed_changes
            - hnsw_disabled
            - recall_below_minimum
            - weak_multi_layer_topology
    AnnNoFallbackDecision:
      type: object
      required:
      - allowed
      - reasons
      properties:
        allowed:
          type: boolean
          description: True only when the explicit no-fallback rollout policy allows this ANN report.
        reasons:
          type: array
          items:
            type: string
            enum:
            - rollout_disabled
            - fallback_enabled
            - slo_not_required
            - not_hnsw_graph_path
            - fallback_occurred
            - fallback_reason_present
            - report_not_production_safe
            - slo_violations_present
            - empty_graph
            - no_eligible_candidates
            - no_results
            - weak_multi_layer_topology
            - recall_missing
            - recall_below_minimum
            - search_policy_below_rollout_minimum
    AnnEvaluationResponse:
      type: object
      required:
      - available
      - reason
      - ann_report
      - exact_top_k
      - ann_top_k
      - overlap_count
      - recall_q16
      properties:
        available:
          type: boolean
        reason:
          anyOf:
          - type: string
            enum:
            - requires_persisted_checkpoint_without_wal_tail
          - type: 'null'
        ann_report:
          anyOf:
          - $ref: '#/components/schemas/AnnSearchReport'
          - type: 'null'
        no_fallback_decision:
          anyOf:
          - $ref: '#/components/schemas/AnnNoFallbackDecision'
          - type: 'null'
        exact_top_k:
          type: array
          items:
            type: integer
        ann_top_k:
          type: array
          items:
            type: integer
        overlap_count:
          type: integer
        recall_q16:
          type: integer
          minimum: 0
          maximum: 65535
...